# club_cms/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Club CMS API",
        "description": "Documentación de la API para el Club CMS",
        "contact": {
            "email": "soporte@clubcms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Base URL para todos los endpoints"
        }
    ],
    "paths": {
...
        "/sports": {
            "get": {
                "tags": [
                    "Sports"
                ],
                "summary": "Listar deportes",
                "description": "Obtiene una lista paginada de todos los deportes. Accesible para usuarios y administradores autenticados.",
                "operationId": "8171a98b44afe5a0817781eed04221fb",
                "responses": {
                    "200": {
                        "description": "Listado de deportes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sports": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SportResource"
                                            }
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_sports": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay deportes disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los deportes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sports"
                ],
                "summary": "Crear un nuevo deporte (Solo Admins)",
                "description": "Crea un nuevo deporte en el sistema. Requiere rol de administrador.",
                "operationId": "985d0ccb04225afb2d090fb71cb7d059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deporte creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sport": {
                                            "$ref": "#/components/schemas/SportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el deporte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sports/{id}": {
            "get": {
                "tags": [
                    "Sports"
                ],
                "summary": "Mostrar detalles de un deporte",
                "description": "Obtiene los detalles de un deporte específico. Accesible para usuarios y administradores autenticados.",
                "operationId": "b8a8723a0a73b793da9a97c2e538f64f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del deporte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sport": {
                                            "$ref": "#/components/schemas/SportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el deporte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sports"
                ],
                "summary": "Actualizar un deporte existente (Solo Admins)",
                "description": "Actualiza los datos de un deporte existente. Requiere rol de administrador.",
                "operationId": "08935a01fafba8708378e15d0441038a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sport": {
                                            "$ref": "#/components/schemas/SportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el deporte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sports"
                ],
                "summary": "Eliminar un deporte (Solo Admins)",
                "description": "Elimina un deporte del sistema. Requiere rol de administrador.",
                "operationId": "390abb277eb6fb39983e919bb33015cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deporte eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deporte eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar el deporte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SportRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fútbol"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción del deporte"
                    }
                },
                "type": "object"
            },
            "SportResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fútbol"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción del deporte"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "description": "Respuesta estándar para errores internos del servidor",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje de error seguro que no expone detalles técnicos",
                        "type": "string",
                        "example": "Error interno del servidor al procesar la solicitud."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "description": "Respuesta cuando falla la validación de los datos de la petición",
                "properties": {
                    "ok": {
                        "description": "Indica que la operación falló",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje de error general",
                        "type": "string",
                        "example": "Error de validación"
                    },
                    "errors": {
                        "description": "Listado de errores de validación por campo",
                        "type": "object",
                        "example": {
                            "email": [
                                "El correo electrónico ya está registrado por otro miembro."
                            ],
                            "member_id": [
                                "El miembro ya tiene el máximo de 3 reservas permitidas para esta fecha."
                            ],
                            "court_id": [
                                "Ya existe una reserva para esta cancha en la misma fecha y hora."
                            ],
                            "date": [
                                "La fecha no puede ser anterior a hoy."
                            ],
                            "hour": [
                                "El horario debe estar entre las 08:00 y las 21:00."
                            ],
                            "name": [
                                "El campo nombre es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}